refactor(mesh): separate create and edit collection routes
