+++ packages/mesh/app/actions/collection.ts
@@ -15,7 +15,7 @@ import queryGraphQL, { mutateGraphQL } from '@/utils/fetch-graphql'
 import { fetchRestfulPost } from '@/utils/fetch-restful'
 import { getLogTraceObjectFromHeaders } from '@/utils/log'
 
-import type { CollectionFormat } from '../collection/(mutate)/_types/edit-collection'
+import type { CollectionFormat } from '../collection/(mutate)/_types/collection'
 
 export async function getCollection({
   collectionId,
