+++ packages/mesh/app/collection/(mutate)/(edit)/(mobile)/edit-stories/[collectionId]/page.tsx
@@ -1,0 +1,5 @@
+export default function EditCollectionStoriesPage() {
+  return (
+    <main className="flex w-full max-w-[theme(width.maxContent)] grow flex-col"></main>
+  )
+}
