+++ packages/mesh/app/collection/(mutate)/(edit)/(mobile)/edit-title/[collectionId]/page.tsx
@@ -1,0 +1,5 @@
+export default function EditCollectionTitlePage() {
+  return (
+    <main className="flex w-full max-w-[theme(width.maxContent)] grow flex-col"></main>
+  )
+}
