--- packages/mesh/app/collection/(mutate)/[id]/edit/layout.tsx
@@ -1,16 +1,0 @@
-import LayoutTemplate from '@/components/layout-template'
-
-export default function NewCollectionLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return (
-    <LayoutTemplate
-      type="collection"
-      mobileNavigation={{ leftButtons: [], title: '', rightButtons: [] }}
-    >
-      {children}
-    </LayoutTemplate>
-  )
-}
