--- packages/mesh/app/collection/(mutate)/[id]/edit/page.tsx
@@ -1,7 +1,0 @@
-export default function NewCollectionPage() {
-  return (
-    <main className="w-full max-w-[theme(width.maxContent)] grow">
-      Edit Collection
-    </main>
-  )
-}
