--- packages/mesh/app/collection/(mutate)/_components/desktop/collection-picks/step-two-sort-stories.tsx
@@ -1,5 +1,0 @@
-import SortStories from '../../sort-stories'
-
-export default function DesktopStep2SortStories() {
-  return <SortStories />
-}
