--- packages/mesh/app/collection/(mutate)/_components/desktop/desktop-edit-collection.tsx
@@ -1,11 +1,0 @@
-import DesktopCollectionPicks from './collection-picks'
-import DesktopMainAction from './main-action'
-
-export default function DesktopEditCollection() {
-  return (
-    <div className="hidden grow flex-col lg:flex">
-      <DesktopMainAction />
-      <DesktopCollectionPicks />
-    </div>
-  )
-}
