--- packages/mesh/app/collection/(mutate)/_components/mobile/new-collection/step-four-sort-stories.tsx
@@ -1,15 +1,0 @@
-import SortStories from '../../sort-stories'
-import TabletGoNextButton from '../../tablet/tablet-go-next-button'
-
-export default function MobileStep4SortStories() {
-  return (
-    <div className="flex grow flex-col pl-2 pr-5 sm:px-5 md:px-[70px] lg:pl-10 lg:pr-0">
-      <SortStories />
-      <div className="fixed inset-x-0 bottom-0 hidden items-center justify-center bg-white py-4 sm:flex lg:hidden">
-        <div className="w-[335px]">
-          <TabletGoNextButton />
-        </div>
-      </div>
-    </div>
-  )
-}
