--- packages/mesh/app/collection/(mutate)/_components/tablet/tablet-navigation.tsx
@@ -1,13 +1,0 @@
-import MobileGoBackButton from '../mobile/mobile-go-back-button'
-import MobileTitle from '../mobile/mobile-title'
-
-export default function TabletNavigation() {
-  return (
-    <div className="hidden items-center gap-5 px-5 py-4 sm:flex md:px-[70px] lg:hidden">
-      <MobileGoBackButton />
-      <h2 className="list-title text-primary-800">
-        <MobileTitle />
-      </h2>
-    </div>
-  )
-}
