+++ packages/mesh/app/collection/(mutate)/_utils/merge-pick-and-bookmark-stories.ts
@@ -1,4 +1,4 @@
-import type { PickOrBookmark } from '../_types/edit-collection'
+import type { PickOrBookmark } from '../_types/collection'
 
 export function mergePickAndBookmarkStories(
   picks: PickOrBookmark[],
