+++ packages/mesh/app/collection/(mutate)/new/_components/desktop/collection-picks/step-one-full-edit.tsx
@@ -1,0 +1,35 @@
+import PickStories from '@/app/collection/(mutate)/_components/pick-stories'
+import PickStoryCard from '@/app/collection/(mutate)/_components/pick-story-card'
+import type { CollectionPickStory } from '@/app/collection/(mutate)/_types/collection'
+
+export default function DesktopStep1FullEdit({
+  fixedStory,
+}: {
+  fixedStory: CollectionPickStory | null
+}) {
+  return !fixedStory ? (
+    <Step1PickStories />
+  ) : (
+    <Step1FixedStory fixedStory={fixedStory} />
+  )
+}
+
+const Step1PickStories = () => {
+  return <PickStories />
+}
+
+const Step1FixedStory = ({
+  fixedStory,
+}: {
+  fixedStory: CollectionPickStory | null
+}) => {
+  if (!fixedStory) return null
+  return (
+    <PickStoryCard
+      key={fixedStory.id}
+      isPicked={true}
+      onClick={() => {}}
+      story={fixedStory}
+    />
+  )
+}
