+++ packages/mesh/app/collection/(mutate)/new/_components/desktop/collection-picks/step-two-sort-stories.tsx
@@ -1,0 +1,5 @@
+import SortStories from '@/app/collection/(mutate)/_components/sort-stories'
+
+export default function DesktopStep2SortStories() {
+  return <SortStories />
+}
