+++ packages/mesh/app/collection/(mutate)/new/_components/desktop/create-collection.tsx
@@ -1,0 +1,11 @@
+import DesktopCollectionPicks from './collection-picks'
+import DesktopMainAction from './main-action'
+
+export default function DesktopCreateCollection() {
+  return (
+    <div className="hidden grow flex-col lg:flex">
+      <DesktopMainAction />
+      <DesktopCollectionPicks />
+    </div>
+  )
+}
