+++ packages/mesh/app/collection/(mutate)/new/_components/mobile/steps/step-four-sort-stories.tsx
@@ -1,0 +1,16 @@
+import SortStories from '@/app/collection/(mutate)/_components/sort-stories'
+
+import TabletGoNextButton from '../../tablet/go-next-button'
+
+export default function MobileStep4SortStories() {
+  return (
+    <div className="flex grow flex-col pl-2 pr-5 sm:px-5 md:px-[70px] lg:pl-10 lg:pr-0">
+      <SortStories />
+      <div className="fixed inset-x-0 bottom-0 hidden items-center justify-center bg-white py-4 sm:flex lg:hidden">
+        <div className="w-[335px]">
+          <TabletGoNextButton />
+        </div>
+      </div>
+    </div>
+  )
+}
