+++ packages/mesh/app/collection/(mutate)/new/_components/mobile/steps/step-two-set-title.tsx
@@ -1,0 +1,18 @@
+import EditHeroImage from '@/app/collection/(mutate)/_components/edit-hero-image'
+import EditTitle from '@/app/collection/(mutate)/_components/edit-title'
+
+import TabletGoNextButton from '../../tablet/go-next-button'
+
+export default function MobileStep2SetTitle() {
+  return (
+    <div className="flex flex-col gap-8">
+      <EditHeroImage />
+      <EditTitle autoFocus={true} />
+      <div className="hidden items-center justify-center pt-6 sm:flex">
+        <div className="w-[295px]">
+          <TabletGoNextButton />
+        </div>
+      </div>
+    </div>
+  )
+}
