+++ packages/mesh/app/collection/(mutate)/new/_types/create-collection.ts
@@ -1,0 +1,11 @@
+export enum MobielCreateCollectionStep {
+  Step1SelectStories = 'mobile-step-1-select-stories',
+  Step2SetTitle = 'mobile-step-2-set-title',
+  Step3SetSummary = 'mobile-step-3-set-summary',
+  Step4SortStories = 'mobile-step-4-sort-stoires',
+}
+
+export enum DesktopCreateCollectionStep {
+  Step1EditAll = 'desktop-step-1-edit-all',
+  Step2SortStories = 'desktop-step-2-sort-stories',
+}
