+++ packages/mesh/utils/cross-page-create-collection.ts
@@ -1,4 +1,4 @@
-import type { CollectionPickStory } from '@/app/collection/(mutate)/_types/edit-collection'
+import type { CollectionPickStory } from '@/app/collection/(mutate)/_types/collection'
 import { MINUTE } from '@/constants/time-unit'
 
 type CreateCollectionStory = {
